Bug 1201997 - Part 4 - Change size and offset variables to size_t.r=froydn
UltraBlame original commit: 9093e927a2474f41c30d2f5239b98e00b6cf6951

+++ netwerk/protocol/http/Http2Session.cpp
@@ -1098,8 +1098,8 @@ Http2Session::CleanupStream(uint32_t aID, nsresult aResult, errorType aResetCode
 
 static void RemoveStreamFromQueue(Http2Stream *aStream, nsDeque &queue)
 {
-  uint32_t size = queue.GetSize();
-  for (uint32_t count = 0; count < size; ++count) {
+  size_t size = queue.GetSize();
+  for (size_t count = 0; count < size; ++count) {
     Http2Stream *stream = static_cast<Http2Stream *>(queue.PopFront());
     if (stream != aStream)
       queue.Push(stream);
@@ -1846,8 +1846,8 @@ Http2Session::RecvGoAway(Http2Session *self)
   self->mStreamTransactionHash.Enumerate(GoAwayEnumerator, self);
 
   
-  uint32_t size = self->mGoAwayStreamsToRestart.GetSize();
-  for (uint32_t count = 0; count < size; ++count) {
+  size_t size = self->mGoAwayStreamsToRestart.GetSize();
+  for (size_t count = 0; count < size; ++count) {
     Http2Stream *stream =
       static_cast<Http2Stream *>(self->mGoAwayStreamsToRestart.PopFront());
 
@@ -1864,7 +1864,7 @@ Http2Session::RecvGoAway(Http2Session *self)
   
   
   size = self->mQueuedStreams.GetSize();
-  for (uint32_t count = 0; count < size; ++count) {
+  for (size_t count = 0; count < size; ++count) {
     Http2Stream *stream =
       static_cast<Http2Stream *>(self->mQueuedStreams.PopFront());
     MOZ_ASSERT(stream->Queued());
